Nim: Use `csize_t` in appropriate places to match EVMC API
EVMC uses `size_t` in a number of places where we were using Nim `uint` before
this patch.  Change the Nim version to use Nim `csize_t`, which is C `size_t`.

Nim's `uint` ("same size as a pointer") is _probably_ always the same size and
type in generated C code as `size_t`, but this is not guaranteed.  (It is known
to be false on some old machines we're unlikely to target.)

So this change is _probably_ not necessary for binary compatibility, but it's
better to be sure.

Signed-off-by: Jamie Lokier <jamie@shareable.org>