Nim: Disable "Imported and not used" warning in test program
The Nim example host/VM module is only called via functions exported and
imported at the C level with `.exportc` -> `.importc`.

No symbols are used from the module at the Nim level, so Nim thinks it is
an unused import and warns:

    Warning: imported and not used: 'nim_host' [UnusedImport]

But we must import so the module is compiled into the program for the C
functions to be available, so disable the warning for this one module.

Signed-off-by: Jamie Lokier <jamie@shareable.org>