Nim: Convert `evmc_flags` to Nim bitset
A bitset (`set[enum_type]`) is more Nim style.

We must use `{.size: sizeof(uint32).}` to keep binary compatibility with the
EVMC API.  Nim uses an unsigned 32-bit type for `set[]` when the size is given
like this, so it works.  Without this, flag issues tend to be hidden on
little-endian architectures but there can be junk extra flags passed that may
have unexpected effects.  On big-endian architectures it is essential.

In the test, `{EVMC_STATIC}` is added to flags just to check the type works.
If it wasn't a bitset, that syntax would fail.

Signed-off-by: Jamie Lokier <jamie@shareable.org>