Nim: Convert `evmc_capabilities` to Nim bitset
A bitset (`set[enum_type]`) is more Nim style.

Custom overloads for `incl`, `excl` and `contains` are no longer required
because they come for free with a bitset.

We must use `{.size: sizeof(uint32).}` to keep binary compatibility with the
EVMC API.  Nim uses an unsigned 32-bit type for `set[]` when the size is given
like this, so it works.  Without this, flag issues tend to be hidden on
little-endian architectures but there can be junk extra flags passed that may
have unexpected effects.  On big-endian architectures it is essential.

Signed-off-by: Jamie Lokier <jamie@shareable.org>