Nim: Make `evmc_status_code` a normal enum
This is more Nim style, and aligns better with the layout in the EVMC API
definition for other languages.  The `==` overload is no longer required.

The negative items have to be put out of order compared with other EVMC API
definitions (a Nim requirement), and this might be why a not-really-enum was
used before.  I think the documentation benefit of having the values in place
and using a standard enum type outweighs this.

Signed-off-by: Jamie Lokier <jamie@shareable.org>