Nim: Recognise that target C enums are not always `cint`
EVMC uses C `enum` types as parameters and in structures.

Although these are nearly always the size of C `int`, a few targets default
to `char` (because all the enum values fit).  Just our luck, it's some 32-bit
ARM ABIs (AAPCS and BPAPI); but not Linux, FreeBSD or NetBSD.  It is `char`
on 32-bit ARM Google Fuchsia, which we might realistically encounter, and
some embedded system CPUs.

We'll assume C `int`.  If you run on those 32-bit ARMs and EVMC doesn't work,
this comment and type may prove useful.

Signed-off-by: Jamie Lokier <jamie@shareable.org>