Nim: Compile time verify that target C enums are `int` sized
Check small-range enums have C `int` size, so the definitions in this file
are binary compatible with EVMC API in `evmc.h`.  On almost all targets it's
compatible.  This is to avoid lost time debugging on a target where it isn't.

The negative-size array trick is portable C equivalent to `static_assert`.

Signed-off-by: Jamie Lokier <jamie@shareable.org>