Nim: Fix missing `ptr` in signature of `execute`
Add `ptr` which was missing on the message parameter in `execute` in the
Nim API.  With this change it matches the C header `<evmc/evmc.h>`, where all
parameters larger than 64 bits are passed by address.

This doesn't change the binary compatibility because Nim converts some
parameters to pass by reference anyway, but we were relying on Nim's heuristic.
Adding `ptr` makes it explicit and matches the rest of the API.

Signed-off-by: Jamie Lokier <jamie@shareable.org>