Nim: Use `ptr` in `release` for API consistency
Nearly all API functions use `ptr`, so change `release` to use it as well.

Add a helper `res.destroy()` to the support API in `evmc_nim.nim` and use it in
the test program.  It's a destructor so this is a fine name.  (We can't call it
`res.release()` because that name conflicts with the object field.)

In the support API, remember to check `res.release` for nil before calling it.
Nil doesn't come up in our test program, but a real EVM might use it.

Signed-off-by: Jamie Lokier <jamie@shareable.org>