Nim: Use `var` instead of `ptr` for arguments that are pass-by-ref
Use `var` instead of `ptr` in the Nim EVMC API `import evmc/evmc` because all
users of the API are already using `var`.

`var` is more convenient to work with, and API users use casts to ignore type
mismatches.  Therefore, change the API itself to use `var`, so it's better to
use directly, casts are not needed, and proper type checking can be used.

Another reason to do this is it brings Nim EVMC a step closer to the types
actually used by API users, so eventually it won't be necessary to `cast[..]`
function types, and proper type-checking can take place.

This is technically a breaking API change, but since the API users use casts to
ignore type mismatches at the moment, it shouldn't break them.

Note: This is actually a stepping stone; we'll change to `{.byref.}` in a
future patch, to represent `const type *` (read-only reference) which `var`
does not.  This `var` version is useful for testing and in case we need to
revert to it if there's a problem with `{.byref.}`.

Passing by reference (address) in Nim + C
-----------------------------------------

When translating a C binary interface to Nim, we can represent function
arguments which are pointers in C as any of these in Nim:

- `ptr type`
- `var type`
- `type` assuming Nim's size heuristic picks pass by reference
- `type` with `{.byref.}` on the type definition

When passing by-reference, so far `ptr type` was used in Nim EVMC in `evmc.nim`
in most places to define the API.  It was a more direct transalation of the C
file, and `evmc_nim.nim` made extensive use of `unsafeAddr` to make calls from
the VM side.

However, the example host in `nim_host.nim` defined its functions arguments
with `var type`, so the functions didn't match the type signature defined in
the API.  It worked because the function pointers were cast (so Nim did not
check the types match), and because `ptr type` and `var type` are
binary-compatible in Nim, even though they are not type-compatible.

Nimbus Eth1's EVMC support also defines functions with `var type` the same way
as `nim_host.nim`, and even its interface file uses `var type`.

Since everything already relies on `var type` being a pointer argument
"silently dereferenced" in Nim, switch the Nim EVMC API in `evmc.nim` to also
use `var type`, to match the type signatures of implementation functions.

This is less clunky to use.  Implementations don't need to dereference pointers, and
there's no need for `unsafeAddr` all over the place; these are removed from
`evmc_nim.nim`.

Which types and arguments are changed
-------------------------------------

The following types are candidates for `var` as function arguments.  These are
conceptually value types, which are passed by-const-reference for efficiency in
the EVMC C `<evmc/evmc.h>` standardised binary interface.  Therefore we also
pass them using `var` when they appear as an argument.  Values of these types
are returned by-value in the current binary interface:

- `evmc_address`
- `evmc_bytes32`
- `evmc_uint256be`
- `evmc_message`
- `evmc_result`
- `evmc_tx_context`

Not all these types appear in arguments currently, but if they did we would
make them `var`.

`evmc_result` is special.  It has a destructor but is returned by copy.  This
implies data which is safe to copy again, and implies the destructor doesn't
care about the object address, only the contents.  Yet it is only safe to call
the destructor on one of the copies, so this is not like an OOP destructor.

Things kept the same
--------------------

Variable size array arguments: `ptr byte` arrays and the `topics` array.

These arguments representing opaque handles are kept as `ptr`:

- `evmc_host_context` (`distinct pointer` is the type)
- `ptr evmc_vm`

These are opaque handles on one side of the EVMC host/VM divide.  They are kept
as `ptr` to say on the opaque side, "be careful" and "don't mess with this".

Even though `evmc_vm` is a Nim `object`, it is the "base object" of a
potentially larger object.  It must not be copied ("[object slicing]
(https://en.wikipedia.org/wiki/Object_slicing)") and the original address must
always be used.

When objects like these are passed as `var`, saving them requires conversion to
`ptr` anyway with `unsafeAddr`.  Passing the saved value to another function
requires dereferencing with `[]` even though we know the `[]` is cancelled by
the receiving `var` and it's not really dereferenced.  None of that is helpful
or clean; better to keep this sort of handle as `ptr`.

In regular Nim we would use `ref` for such types, but not with these because
these are opaque pointers from C.  `ref` is only for values allocated by Nim.

- `ptr emvc_host_interface`

Perhaps surprisingly, this is plain old data, and can be safely copied even if
it's part of a larger object.  Its address is not important either.  The reason
for sticking with `ptr` is because the receiver normally stores the address
without copying the contents, even though it would be safe to copy.  If it was
`var` the receiver would use `unsafeAddr` to make a `ptr` anyway.

Signed-off-by: Jamie Lokier <jamie@shareable.org>