Nim: Make the test host use exactly the type signatures in the API
Before this patch, the host functions have first argument `HostContext` which
doesn't match the EVMC API type `evmc_host_context.  These functions are
type-incompatible with the dispatch table, but binary compatible in practice,
so they are casted with `cast[..]` to store in the dispatch table.

This was a convenient way to write, but just accepting `evmc_host_context` as
the first argument and immediately converting it to `HostContext` is simple as
well, and has the advantage of exactly matching the type in the EVMC API, so
functions can be stored with types checked properly.

Signed-off-by: Jamie Lokier <jamie@shareable.org>